chore: move register-strategy into runtime
