+++ core/base/index.ts
@@ -8,4 +8,8 @@ export * from "./indexer-helpers.js";
 
 export * from "./version.js";
 
-export * from "./register-compact-strategy.js";
+import "./compact-strategies.js";
+
+export { getCompactStrategy, registerCompactStrategy, getCompactStrategyThrow } from "@fireproof/core-runtime";
+
+export { registerStoreProtocol } from "@fireproof/core-blockstore";
